Get rid of sru module
